+++ packages/commercetools/theme/pages/Checkout/Shipping.vue
@@ -6,7 +6,7 @@
       class="sf-heading--left sf-heading--no-underline title"
     />
     <ValidationObserver v-slot="{ handleSubmit, dirty, reset }">
-      <form @submit.prevent="handleSubmit(dirty || addressIsModified ? handleShippingAddressSubmit(reset) : handleShippingMethodSubmit(reset))">
+      <form @submit.prevent="handleSubmit(canContinueToPayment(dirty) ? handleShippingMethodSubmit(reset) : handleShippingAddressSubmit(reset))">
         <UserShippingAddresses
           v-if="isAuthenticated && shippingAddresses && shippingAddresses.length"
           :setAsDefault="setAsDefault"
@@ -299,6 +299,7 @@ export default {
       addressIsModified.value = false;
     };
     const handleShippingMethodSubmit = (reset) => async () => {
+      await setShippingMethod(chosenShippingMethod.value, { save: true });
       reset();
       context.root.$router.push('/checkout/payment');
     };